Fix jacoco classloader issues when collecting coverage
In some cases, when the jacoco runtime is provided via TestKit's
withPluginClasspath, the jacoco agent class ends up in a separate
classloader from the coverage plugin.

We ensure that the coverage plugin is in the same classloader as the
jacoco runtime by passing the coverage plugin jar into the TestKit's
plugin classpath instead of letting the TestKit build resolve it
normally. As a minor side effect, the coverage plugin version no longer
needs to be explicitly specified, and the coverage plugin no longer
needs to be published to the gradle plugin portal.